requirements.txt
colorama==0.4.6
llama-index==0.10.36
llama-index-vector-stores-chroma==0.1.8
python-dotenv==1.0.1
streamlit==1.34.0
